+++ PCbuild/_testcapi.vcxproj.filters
@@ -120,6 +120,9 @@
     <ClCompile Include="..\Modules\_testcapi\frame.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Modules\_testcapi\type.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\PC\python_nt.rc">
